Incorpore la paginación

+++ frontend/src/app/pages/addressesPage.ts
@@ -1,12 +1,15 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
+import { untilComponentDestroyed } from '@w11k/ngx-componentdestroyed';
 
 import { MatTableDataSource } from '@angular/material/table';
+import { MatPaginator } from '@angular/material/paginator';
 
 import { Address } from '@model/address';
-import { AddressService } from '@services/addressesService';
 import { BaseSectionPage } from '@components/pages/BaseSectionPage';
+import { AddressService } from '@services/addressesService';
 import { UtilServices } from '@services/utilServices';
+import { PopupService } from '@services/popupService';
 import { Constants } from '@utils/constants';
 
 @Component({
@@ -21,6 +24,8 @@ export class AddressesPage extends BaseSectionPage implements OnInit {
   dataSource = new MatTableDataSource<Address>; //datasource
   title: string = "Direcciones IP"
 
+  @ViewChild(MatPaginator) paginator: MatPaginator;
+
     /**
    * flag que marca si es una duplicacion
    */
@@ -30,7 +35,8 @@ export class AddressesPage extends BaseSectionPage implements OnInit {
   /**
    * constructor
    */
-  constructor(fb: FormBuilder, private addressesService: AddressService, utilService: UtilServices) {
+  constructor(fb: FormBuilder, private addressesService: AddressService, utilService: UtilServices,
+              private popupService: PopupService ) {
     super(fb, utilService);
     const state = this.utilService.getState();
     this.duplicate=state && state.duplicate;
@@ -38,6 +44,10 @@ export class AddressesPage extends BaseSectionPage implements OnInit {
 
   getTitle() {
     return this.title;
+  }
+
+  ngAfterViewInit(): void {
+    this.dataSource.paginator = this.paginator;
   }
 
   ngAfterViewChecked(): void {
@@ -53,6 +63,11 @@ export class AddressesPage extends BaseSectionPage implements OnInit {
   //==================== Metodos =====================
   clickedRows = new Set<Address[]>();
 
+  applyFilter(event: Event): void {
+    const filterValue = (event.target as HTMLInputElement).value;
+    this.dataSource.filter = filterValue.trim().toLowerCase();
+  }  
+
   getAddresses() {
     this.addressesService.getAddresses().subscribe({
       next: res => {
@@ -63,14 +78,27 @@ export class AddressesPage extends BaseSectionPage implements OnInit {
     
   }
 
-  editAddress(id: string, event: { stopPropagation: () => void; }) {
-    event.stopPropagation();
-    this.addressesService.getAddress(id)
-    .subscribe({
-      next: result => {
-        this.utilService.navigate(Constants.URL.ADDRESS, { edit: true, data: result })
-      }
-    });
+  editAddress(address: Address) {
+    console.log('editAddress', address);
+    this.utilService.navigate(Constants.URL.ADDRESS, { edit: true, data: address })
+      // .pipe((tap) => console.log('Edit Address -> ' + JSON.stringify(data)))
+  
+      // .pipe(untilComponentDestroyed(this))
+      // .subscribe(res => {
+        // if (res) {
+        //   const text: string = `¿confirma los cambios realizados a la dirección IP ?`;
+        //   this.popupService.showConfirmationPopup(text).afterClosed()
+          // this.addressesService.updateAddress(address)
+        // } else {
+        //   this.cancel();
+        // }
+      // })
+    // this.addressesService.getAddress(id)
+    // .subscribe({
+    //   next: result => {
+    //     this.utilService.navigate(Constants.URL.ADDRESS, { edit: true, data: result })
+    //   }
+    // });
       // {
       // next: res => {
       //   if (res._id) {
@@ -84,9 +112,18 @@ export class AddressesPage extends BaseSectionPage implements OnInit {
     // })
   }
 
-  deleteAddress(id: string, $event: MouseEvent) {
+  deleteAddress(id: string) {
     console.log('deleteAddress', id);
-    this.addressesService.deleteAddress(id)
+    const text: string = `¿Está seguro que desea borrar la dirección IP ?`;
+    this.popupService.showConfirmationPopup(text).afterClosed()
+      .pipe(untilComponentDestroyed(this))
+      .subscribe(res => {
+        if (res) {
+          this.addressesService.deleteAddress(id)
+        } else {
+          this.cancel();
+        }
+      })
     
   }
